Fixing several #ifdef issues concerning HAS_HEATED_PLATFORM and adding optional support for I2C-LC-Display
In order to disable the HAS_HEATED_PLATFORM code by setting the #define to 0 (instead of commenting it out), several #ifdef resp. #ifndef statements needed to be changed to #if resp. #if ! statements.

The optional I2C-LC-Display support can be enabled by setting the #define HAS_I2C_LCD to 1 (in Configuration.h). In case HAS_I2C_LCD is enabled, we have a low memory problem. To solve this, the currently unused EEPROM and SERVO handling is automatically disabled when HAS_I2C_LCD is enabled (and vice versa).

The I2C_LCD support within the extruder controller currently supports basic display of states handled by the controller:
- Extruder temperature (current & target)
- Platform temperature (current & target, if enabled/available)
- Extruder-Motor ("Extruding..."/"Idle")