Revar Desmera's I2C LCD and 16-button matrix keypad support
merged with master and uploaded to github by Jordan Miller

google group discussion (scroll down):
http://groups.google.com/group/makerbot/browse_thread/thread/a171e6ad8f7e432

Revar says:
I'm using the I2C bus pins to drive a $40 Modtronix LCD2S serial 20x4 LCD display.
  ( http://www.modtronix.com/index.php?cPath=131_141 )
The Modtronics controller has support for up to a 4x4 keypad, so I just poll the LCD to get keypresses.  I don't have a formal schematic for it, but the hookup was really simple based on the info in the display's dataspecs.
  ( http://www.modtronix.com/products/lcd2s/lcd2sr1_v140.pdf )
I just used individual test jumper leads that slide right onto the pins.  I'd take a photo of the wiring, but I'm in a 2 hour build of a part right now.  :)
I've made a few fixes and improvements to the firmware from what is available for download.  I'll package up the improvements and put them up tomorrow.
        - Revar