Started changing Adam's universal homing script so that it homes X,Y before Z to avoid Z crashes.
Right now it only will home the XY when it receives the order to home all downwards. now it needs to wait and then home z.

perhaps it should also do this when it receives any type of homing command that involves z and another axis in the minus direction?
(Im assuming that the user knew what they were doing when they said home z only. maybe should assume differently?)