Adam got it working! It now saves all three values to EEMPROM!
Aparently the "&" infront of the varible says point to the location of the varible (In memory) and I was passing in for the location of which to write the
data, a point in flash memory (the point in which the varible "offset" is stored.) Adam deleted the & symbol and it now passes in the location IN the varible.

Now to tackle the simultaneous homing problem.