Added to ability to home carefully in both directions.
Still untested. Also need to make the z rise before homing down when homing down.
Also need to have it not wait for the long command to finish. i.e. start a command, tend to other things and just check to see if it's ready
to begin the next step in the script. Maybe should move all complex scripts to a scripts.cc and .h file and handle them just like wait for tools.